Address review: move fitToWidth to utils, simplify guard, poll for zoom
- Move fitToWidth.ts -> frontend/src/utils/pdfZoom.ts to break the
  annotator <-> knowledge_base back-edge import (CLAUDE.md section 6 —
  pure number-in/number-out utilities belong in utils/, depending
  only on constants.ts).
- Collapse the redundant '!naturalPageWidth || naturalPageWidth <= 0'
  null-guard to a single '<= 0' check that covers 0, NaN, and negatives.
- Replace the verbose three-line useEffect comment in PDFPage.tsx with
  a one-line WHY pointer to issue #1736.
- Replace 'waitForTimeout(1500)' in the fit-to-width CT group with a
  'waitForFunction' that polls .zoom-level until the parsed value
  exceeds 100 % (visible evidence, not a fixed sleep).
- Refresh FIT_WIDTH_MARGIN docstring to say '16 px total breathing
  room reserved' (no longer references the misleading 'split between
  the two edges' wording) and to point at the new utils path.
