+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- **`DocumentKnowledgeBase` desktop initial PDF zoom — reserve a horizontal margin so the fit-to-width result never overflows on narrower laptop viewports** (issue #1736; `frontend/src/components/annotator/renderers/pdf/PDFPage.tsx`, `frontend/src/components/knowledge_base/document/document_kb/fitToWidth.ts` (new), `frontend/src/components/knowledge_base/document/layouts/mobile/useMobileFitToWidth.ts`, `frontend/src/assets/configurations/constants.ts`). Verified the desktop initial PDF zoom is already a fit-to-width calculation derived from the measured document-layer width (not a fixed seed), but the calculation reserved no horizontal margin — the rendered page touched both edges of the container, leaving no safety budget for sub-pixel rounding or container chrome and risking a horizontal scrollbar on 1280 px / 1024 px laptops. The desktop calculation now flows through the same shared `computeFitToWidthZoom(naturalPageWidth, containerWidth)` helper used by the mobile fit-to-width hook (issue #1735's mobile work introduced `FIT_WIDTH_MARGIN`), reserving `FIT_WIDTH_MARGIN` (16 px) off the container width and clamping to the canonical `ZOOM_MIN`/`ZOOM_MAX` (50 % / 400 %) — previously the desktop path clamped to the magic numbers `0.3` / `4.0`, drifting from the shared constants. The mobile hook is refactored to call the shared helper so both surfaces stay in lockstep going forward. New unit test `frontend/src/components/knowledge_base/document/document_kb/__tests__/fitToWidth.test.ts` pins the helper's contract; new CT test group `Default PDF zoom fit-to-width — {viewport} (#1736)` in `frontend/tests/DocumentKnowledgeBase.ct.tsx` mounts the full DKB at 1440×900, 1280×720 and 1024×768 and asserts that the rendered PDF canvas width is strictly less than the `#pdf-container` width at every viewport, capturing one documentation screenshot per viewport (`knowledge-base--pdf-default-zoom--1440x900.png`, `--1280x720.png`, `--1024x768.png`).
+- **`DocumentKnowledgeBase` desktop initial PDF zoom — reserve a horizontal margin so the fit-to-width result never overflows on narrower laptop viewports** (issue #1736; `frontend/src/components/annotator/renderers/pdf/PDFPage.tsx`, `frontend/src/utils/pdfZoom.ts` (new), `frontend/src/components/knowledge_base/document/layouts/mobile/useMobileFitToWidth.ts`, `frontend/src/assets/configurations/constants.ts`). Verified the desktop initial PDF zoom is already a fit-to-width calculation derived from the measured document-layer width (not a fixed seed), but the calculation reserved no horizontal margin — the rendered page touched both edges of the container, leaving no safety budget for sub-pixel rounding or container chrome and risking a horizontal scrollbar on 1280 px / 1024 px laptops. The desktop calculation now flows through the same shared `computeFitToWidthZoom(naturalPageWidth, containerWidth)` helper used by the mobile fit-to-width hook (issue #1735's mobile work introduced `FIT_WIDTH_MARGIN`), reserving `FIT_WIDTH_MARGIN` (16 px) off the container width and clamping to the canonical `ZOOM_MIN`/`ZOOM_MAX` (50 % / 400 %) — previously the desktop path clamped to the magic numbers `0.3` / `4.0`, drifting from the shared constants. The mobile hook is refactored to call the shared helper so both surfaces stay in lockstep going forward. The helper lives under `frontend/src/utils/` (pure number-in/number-out, depends only on `constants.ts`) to keep the `annotator` and `knowledge_base` component subtrees free of a back-edge import cycle. New unit test `frontend/src/utils/__tests__/pdfZoom.test.ts` pins the helper's contract; new CT test group `Default PDF zoom fit-to-width — {viewport} (#1736)` in `frontend/tests/DocumentKnowledgeBase.ct.tsx` mounts the full DKB at 1440×900, 1280×720 and 1024×768 and asserts that the rendered PDF canvas width is strictly less than the `#pdf-container` width at every viewport, capturing one documentation screenshot per viewport (`knowledge-base--pdf-default-zoom--1440x900.png`, `--1280x720.png`, `--1024x768.png`).
 
 - **Backend test suite speedup — Phase A3b: portable buildx cache, shared across CI + local dev + prod** (issue #1769). Swaps the GitHub-Actions-only buildx cache backend (`type=gha`, landed in Phase A3 / PR #1774) for a portable registry-backed cache at `ghcr.io/open-source-legal/opencontracts/django-{local,production}-cache:main`. GHA cache only works inside GitHub Actions, so it could not help contributors building from a fresh checkout. The registry cache is pullable from any docker buildx invocation that can reach ghcr.
   - **CI (`.github/workflows/backend.yml`)**: replaces `--cache-from/--cache-to type=gha,scope=django-test` with `type=registry,ref=ghcr.io/.../django-local-cache:main`. Adds `permissions: packages: write` on the pytest job (needed for the `cache-to` push on main merges), a `docker/login-action@v3` step (using the existing `GITHUB_TOKEN` — no new secrets), and gates `--cache-to` to `push` events only so PR runs read the cache but don't write to it. Drops the inline `actions/github-script` env-var step (registry cache doesn't need `ACTIONS_RUNTIME_TOKEN`).
